Use real subnet prefix instead of sm_lid.
This commit was SVN r11749.

+++ ompi/mca/btl/openib/btl_openib.h
@@ -141,7 +141,7 @@ typedef mca_btl_base_recv_reg_t mca_btl_openib_recv_reg_t;
     
 struct mca_btl_openib_port_info_t {
     uint32_t mtu;
-    uint16_t subnet;
+    uint64_t subnet;
 };
 typedef struct mca_btl_openib_port_info_t mca_btl_openib_port_info_t;
